requirements.txt
pytest
black==23.11.0
Flask-SQLAlchemy==3.0.2
flask-restx==1.2.0
pydantic==2.6.4
Flask==2.2.2
Werkzeug==2.2.2
Flask_Cors==4.0.0
Flask-JWT-Extended==4.5.3
psycopg2-binary
python-dotenv
uvicorn==0.29.0
aiosqlite==0.20.0
passlib 1.7.4
types-passlib==1.7.7.12